requirements.txt
amqp==5.0.6
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
celery==5.1.2
certifi==2022.5.18.1
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
cymem==2.0.6
decorator==5.0.9
docs==0.1.0
docx==0.2.4
et-xmlfile==1.1.0
Flask==2.0.0
flower==1.1.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
humanize==4.2.3
idna==3.3
iniconfig==1.1.1
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jsonify==0.5
kombu==5.1.0
langcodes==3.3.0
lxml==4.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
murmurhash==1.0.7
numpy==1.23.0
oauth2client==4.1.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
parso==0.8.2
pathy==0.6.2
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
pika==0.13.0
Pillow==8.3.1
pluggy==1.0.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.19
protobuf==4.21.2
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.4.0/pt_core_news_lg-3.4.0-py3-none-any.whl
ptyprocess==0.7.0
py==1.11.0
PyAMQP==0.1.0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.9.1
PyDrive==1.3.1
Pygments==2.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.28.0
rsa==4.8
six==1.16.0
smart-open==5.2.1
sortedcontainers==2.4.0
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.3
thinc==8.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.0.5
typer==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.1.2
XlsxWriter==3.0.3
